requirements.txt
numpy==1.16.4
pandas==0.24.2
boto3==1.9.96
botocore==1.12.96
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1
django-cleanup==4.0.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
Pillow==5.2.0
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.5
requests==2.19.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.23
whitenoise==4.1.2
